+++ crates/burn-tensor/src/tensor/nn/functional/linear.rs
@@ -1,5 +1,4 @@
-use crate::tensor::Tensor;
-use crate::tensor::backend::Backend;
+use crate::{Tensor, backend::Backend};
 
 /// Applies a linear transformation to the input tensor using the given weight and bias.
 ///
@@ -41,69 +40,5 @@ pub fn linear<B: Backend, const D: usize>(
     match bias {
         Some(bias) => output + bias.unsqueeze(),
         None => output,
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::TestBackend;
-    use burn_tensor::TensorData;
-
-    #[test]
-    fn test_linear_1d() {
-        let weight =
-            Tensor::<TestBackend, 2>::from_data([[1.0, 2.0], [3.0, 4.0]], &Default::default());
-
-        let x = Tensor::<TestBackend, 1>::from_data([1.0, 2.0], &Default::default());
-
-        linear(x.clone(), weight.clone(), None)
-            .into_data()
-            .assert_eq(
-                &TensorData::from([7.0, 10.0]).convert_dtype(x.dtype()),
-                true,
-            );
-    }
-
-    #[test]
-    fn test_linear_forward_no_bias() {
-        let weight =
-            Tensor::<TestBackend, 2>::from_data([[1.0, 2.0], [3.0, 4.0]], &Default::default());
-
-        let x = Tensor::<TestBackend, 3>::from_data(
-            [[[1.0, 2.0], [3.0, 4.0]], [[-1.0, -2.0], [-3.0, -4.0]]],
-            &Default::default(),
-        );
-
-        linear(x.clone(), weight.clone(), None)
-            .into_data()
-            .assert_eq(
-                &TensorData::from([[[7.0, 10.0], [15.0, 22.0]], [[-7.0, -10.0], [-15.0, -22.0]]])
-                    .convert_dtype(x.dtype()),
-                true,
-            );
-    }
-
-    #[test]
-    fn test_linear_forward_with_bias() {
-        let weight =
-            Tensor::<TestBackend, 2>::from_data([[1.0, 2.0], [3.0, 4.0]], &Default::default());
-        let bias = Some(Tensor::<TestBackend, 1>::from_data(
-            [1.0, -1.0],
-            &Default::default(),
-        ));
-
-        let x = Tensor::<TestBackend, 3>::from_data(
-            [[[1.0, 2.0], [3.0, 4.0]], [[-1.0, -2.0], [-3.0, -4.0]]],
-            &Default::default(),
-        );
-
-        linear(x.clone(), weight.clone(), bias.clone())
-            .into_data()
-            .assert_eq(
-                &TensorData::from([[[8.0, 9.0], [16.0, 21.0]], [[-6.0, -11.0], [-14.0, -23.0]]])
-                    .convert_dtype(x.dtype()),
-                true,
-            );
     }
 }
