+++ crates/burn-tensor/src/tests/mod.rs
@@ -3,6 +3,7 @@ mod clone_invariance;
 mod grid;
 mod linalg;
 mod module;
+mod nn;
 mod ops;
 mod primitive;
 mod quantization;
@@ -155,6 +156,9 @@ macro_rules! testgen_with_float_param {
         burn_tensor::testgen_log_sigmoid!();
         burn_tensor::testgen_silu!();
         burn_tensor::testgen_tanh_activation!();
+
+        // test nn.functional
+        burn_tensor::testgen_nn_fn_vector_norm!();
 
         // test grid
         burn_tensor::testgen_meshgrid!();
