+++ crates/burn-tensor/src/tests/nn/functional/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod vector_norm;
