+++ crates/burn-tensor/src/tests/nn/mod.rs
@@ -1,0 +1,1 @@
+pub(crate) mod functional;
